Move Direction enum to new enums file in manifest-types
