+++ src/planning/strategies/match-recipe-by-verb.ts
@@ -16,9 +16,8 @@ import {Particle} from '../../runtime/recipe/particle.js';
 import {Recipe} from '../../runtime/recipe/recipe.js';
 import {StrategizerWalker, Strategy} from '../strategizer.js';
 import {GenerateParams, Descendant} from '../../runtime/recipe/walker.js';
-import {Direction} from '../../runtime/manifest-ast-nodes.js';
+import {Direction} from '../../runtime/manifest-types/enums.js';
 import {Type} from '../../runtime/type.js';
-import {HandleConnection} from '../../runtime/recipe/handle-connection.js';
 import {Dictionary} from '../../runtime/hot.js';
 
 // This strategy substitutes '&verb' declarations with recipes,
