+++ src/runtime/manifest-types/particle-spec.ts
@@ -10,7 +10,8 @@
 
 import {assert} from '../../platform/assert-web.js';
 import {Modality} from '../modality.js';
-import {Direction, SlotDirection, ClaimStatement, CheckStatement} from '../manifest-ast-nodes.js';
+import {SlotDirection, ClaimStatement, CheckStatement} from '../manifest-ast-nodes.js';
+import {Direction} from './enums.js';
 import {TypeChecker} from '../recipe/type-checker.js';
 import {Schema} from '../schema.js';
 import {InterfaceType, SlotType, Type, TypeLiteral, TypeVariableInfo} from '../type.js';
