+++ src/runtime/recipe/connection-constraint.ts
@@ -11,7 +11,8 @@
 import {assert} from '../../platform/assert-web.js';
 import {ParticleSpec} from '../manifest-types/particle-spec.js';
 
-import {Direction, RELAXATION_KEYWORD} from '../manifest-ast-nodes.js';
+import {RELAXATION_KEYWORD} from '../manifest-ast-nodes.js';
+import {Direction} from '../manifest-types/enums.js';
 import {Handle} from './handle.js';
 import {Comparable, compareArrays, compareComparables, compareStrings, compareBools} from './comparable.js';
 import {Recipe, RecipeComponent, CloneMap, ToStringOptions} from './recipe.js';
