+++ src/runtime/type.ts
@@ -21,6 +21,7 @@ import {ParticleSpec} from './manifest-types/particle-spec.js';
 import {Refinement} from './refiner.js';
 import {AnnotationRef} from './recipe/annotation.js';
 import {ManifestStringBuilder} from './manifest-string-builder.js';
+import { Direction } from './manifest-types/enums.js';
 
 export interface TypeLiteral extends Literal {
   tag: string;
@@ -873,7 +874,7 @@ export class TupleType extends Type {
 export interface HandleConnection {
   type: Type;
   name?: string|TypeVariable;
-  direction?: AstNode.Direction; // TODO make required
+  direction?: Direction; // TODO make required
 }
 
 // TODO(lindner) only tests use optional props
