+++ src/tools/schema2kotlin.ts
@@ -14,7 +14,7 @@ import {generateConnectionSpecType} from './kotlin-type-generator.js';
 import {HandleConnectionSpec, ParticleSpec} from '../runtime/manifest-types/particle-spec.js';
 import {CollectionType, EntityType, Type, TypeVariable} from '../runtime/type.js';
 import {KotlinGenerationUtils} from './kotlin-generation-utils.js';
-import {Direction} from '../runtime/manifest-ast-nodes.js';
+import {Direction} from '../runtime/manifest-types/enums.js';
 import {KotlinEntityGenerator} from './kotlin-entity-generator.js';
 
 // TODO: use the type lattice to generate interfaces
